+++ packages/core/contracts/test/cli/excludes/Abstract1.sol
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.9;
+
+abstract contract Abstract1 {
+  uint256 public immutable x;
+
+  constructor(uint256 _x) {
+    x = _x;
+  }
+}